# Stabilize benchmark result of `BenchmarkValueEqual`